chore: refactor db seed
Consolidate database seeding logic into db.ts
